// src/common/enums/service-type.enum.ts
export enum ServiceType {
  WEB_DEVELOPMENT = 'web_development',
  MOBILE_DEVELOPMENT = 'mobile_development',
  UI_UX_DESIGN = 'ui_ux_design',
  DIGITAL_MARKETING = 'digital_marketing',
  DATA_ANALYTICS = 'data_analytics',
  CLOUD_SERVICES = 'cloud_services',
  CYBERSECURITY = 'cybersecurity',
  AI_ML = 'ai_ml',
  BLOCKCHAIN = 'blockchain',
  CONSULTING = 'consulting'
}